+++ apm-protocol/apm-network/src/main/java/org/apache/skywalking/apm/network/trace/component/ComponentsDefine.java
@@ -213,5 +213,7 @@ public class ComponentsDefine {
 
     public static final OfficialComponent FASTJSON = new OfficialComponent(117, "Fastjson");
 
-    public static final OfficialComponent CLICKHOUSE_JDBC_DRIVER = new OfficialComponent(118, "ClickHouse-jdbc-driver");
+    public static final OfficialComponent JACKSON = new OfficialComponent(118, "Jackson");
+
+    public static final OfficialComponent CLICKHOUSE_JDBC_DRIVER = new OfficialComponent(119, "ClickHouse-jdbc-driver");
 }